LayoutTests/editing/execCommand/indent-nested-lists-5-expected.txt

EDITING DELEGATE: shouldBeginEditingInDOMRange:range from 0 of OL > DIV > BODY > HTML > #document to 9 of OL > DIV > BODY > HTML > #document
EDITING DELEGATE: webViewDidBeginEditing:WebViewDidBeginEditingNotification
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: shouldChangeSelectedDOMRange:(null) toDOMRange:range from 4 of #text > LI > UL > OL > DIV > BODY > HTML > #document to 4 of #text > LI > UL > OL > DIV > BODY > HTML > #document affinity:NSSelectionAffinityDownstream stillSelecting:FALSE
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChange:WebViewDidChangeNotification
This tests indenting "three". You should see 1, 2, and then a bullet before numbers "two", "three", and "four" respectively.
one
two
three
four
Before: <li>one</li> <ol><li>two</li></ol> <li id="test">three</li> <ul><li>four</li></ul>
After: <li>one</li> <ol><li>two</li><li id="test">three</li></ol> <ul><li>four</li></ul>
